Have NotifyMemberShutdownOp registered on live operations registry
Some refactoring of operation base classes
Fixes #1094
