+++ hazelcast-jet-core/src/main/java/com/hazelcast/jet/impl/operation/NotifyMemberShutdownOperation.java
@@ -18,8 +18,6 @@ package com.hazelcast.jet.impl.operation;
 
 import com.hazelcast.jet.impl.JetService;
 import com.hazelcast.jet.impl.execution.init.JetInitDataSerializerHook;
-import com.hazelcast.nio.serialization.IdentifiedDataSerializable;
-import com.hazelcast.spi.Operation;
 
 import java.util.concurrent.CompletableFuture;
 
@@ -30,14 +28,13 @@ import static com.hazelcast.jet.impl.util.ExceptionUtil.withTryCatch;
  * caller is about to shut down. The master should request termination of all
  * jobs running on caller and then the caller will actually shut down.
  */
-public class NotifyMemberShutdownOperation extends Operation implements IdentifiedDataSerializable {
+public class NotifyMemberShutdownOperation extends AsyncOperation {
 
     public NotifyMemberShutdownOperation() {
-        setWaitTimeout(0);
     }
 
     @Override
-    public void run() {
+    protected void doRun() {
         JetService service = getService();
         CompletableFuture<Void> future = service.getJobCoordinationService().addShuttingDownMember(getCallerUuid());
         future.whenComplete(withTryCatch(getLogger(), (r, e) -> sendResponse(null)));
@@ -46,11 +43,6 @@ public class NotifyMemberShutdownOperation extends Operation implements Identifi
     @Override
     public final int getFactoryId() {
         return JetInitDataSerializerHook.FACTORY_ID;
-    }
-
-    @Override
-    public boolean returnsResponse() {
-        return false;
     }
 
     @Override
